Refactoring
Move a bunch of shared things to util in prep for more features.
